WIP Update Spec to 3.3.0
The Spec jump from 2.0.0 to 3.3.0 introduced some
breakage:

There is now a Map type that needs to be modeled
correctly; for now I just made it compile with no
expectation that it works correctly.

The newer spec also has a couple of property types
which are missing a Documentation field, so a
simple method to patch the Spec json was added.
We will need this anyway since the Spec is
incomplete (i.e. `AWS::Backup` missing from the
Spec as of this commit).

There is also now an `Alexa::ASK::Skill` which
breaks assumptions about naming that need to be
fixed; for now we just lump it into the `cfn::aws`
module.

The `AWS::Batch` resources changed enough to break
the tests, but the solution is not
straight-forward due to borrowing rules.